requirements.txt
asgiref==3.7.2
Brotli==1.0.9
channels==4.0.0
channels-redis==4.0.0
click==8.1.3
chromedriver==2.24.1
coverage== 6.5.0
daphne==4.0.0
crispy-bootstrap5==0.7
distro==1.9.0
dj-database-url==2.1.0
django==5.0
django-autoslug==1.9.9
django-cors-headers==4.3.1
django-crispy-forms==1.14.0
djangorestframework==3.14
django-bootstrap5==23.3
dj-database-url==2.1.0
drf-spectacular==0.27.1
factory-boy==3.3.0
Faker==21.0.0
gunicorn==21.2.0
Markdown==3.4.1
markdown-it-py==2.1.0
Markups==4.0.0
MarkupSafe==2.1.2
Pillow==9.4.0
pytest==7.2.1
psycopg2-binary==2.9.9
pytest-cov==4.1.0
pytest-factoryboy==2.5.1
python-dotenv==1.0.0
pytest-django==4.8.0
pytest-factoryboy==2.5.1
pytz==2024.1
redis==5.0.1
regex==2022.10.31
render==1.0.0
requests==2.28.1
selenium==4.16.0
six==1.16.0
sqlparse==0.4.3
websockets==10.4
websocket-client==1.4.2
whitenoise==6.6.0
